+++ src/componentes/BtnV.js
@@ -1,17 +1,15 @@
 import { Link } from "react-router-dom";
 import React from "react";
-import './css/style.css'
-function BtnV(){
+import './css/Style.css';
+
+export default function BtnV(){
     return(
-        <div className="contenedorbotonvolver">
-        <div className="botonvolver">
-         <Link to= "/">    <p> VOLVER </p>  </Link>
-        </div>
-        </div>
+        <button className="contenedorbotonvolver">
+            <Link to= "/">    <p> VOLVER </p>  </Link>
+        </button>
     );
 
 }
-export default BtnV;
