+++ src/componentes/Juego.js
@@ -1,19 +1,29 @@
-import React from "react";
-import './css/style.css'
+import React, { useState } from "react";
+import './css/Style.css';
 import BtnV from "./BtnV.js";
+
 function Juego (props){
+    const word = "GRACIOSO";
+    const letras = ["A", "B", "C", "D", "E", "F", "G",
+        "H", "I", "J", "K", "L", "M", "N", "O", "P", "Q", "R",
+        "S", "T", "U", "V", "W", "X", "Y", "Z"];
+    const [correctGuesses, setCorrectGuesses] = useState([])    
+    const maskedWord = word.split('').map(letter => 
+    correctGuesses.includes(letter) ? letter : "_").join(" ");
 
 
 return(
-
-
-    
-    <div>
-
-<BtnV/>
-  
-
-    </div>
+        <div>
+            <p>{maskedWord}</p>
+            {letras.map((letra, index) => 
+            <button key={index} onClick={() => {
+                if (word.includes(letra)) {
+                    setCorrectGuesses([...correctGuesses, letra])
+                } 
+            }}>{letra}</button>)}
+            {!maskedWord.includes("_") && <p>¡Felicidades! Has ganado</p>}
+        <BtnV/>
+        </div>
 )
 
 }
