+++ src/componentes/css/Style.css
@@ -98,9 +98,7 @@ body
 
    .contenedorbotonhomes{
 
-    display: grid;
-
-    
+    display: grid;    
     margin-left: 39%;
     gap: 40px;
     text-decoration: none;
@@ -122,6 +120,7 @@ body
     border: none;
     cursor: pointer;
     font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif;
+    font-size: 3em;
   }
   
 
@@ -209,9 +208,6 @@ body
 
 
 .botonvolver{
-
-
-
   align-items: center;
   justify-content: center;
   width: 34%;
@@ -223,7 +219,7 @@ body
   cursor: pointer;
   font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif;
   text-align: center;
- 
+  size: 200; 
 }
