--- src/componentes/estilos/Style.css
@@ -1,248 +1,0 @@
-.icono
-{  
-    margin-top: -1%;
-    margin-bottom: 5%;
-    margin-left: 2%;
-
-}
-
-
-body
-{
-
- background-image: linear-gradient(  135deg   ,#692d92 ,#1d76b1,#88acb1,#7277b6,#5fa2ad ,#3c802f ,#80462f,#80422f);
- margin-top: 3.5%;
- margin-bottom: 4%;
- animation: fanimado 25s infinite;
- background-size: 500%;
-}
-
-
-@keyframes fanimado{
-
-0%{background-position: 0% 150% ;
-
-}
-
-50%{background-position: 100% 50% ;
-
-}
-
-100%{background-position: 0% 50% ;
-
-}
-
-}
-
-
-  .contenedorboton
-  {
-    display: grid;
-    margin-top: 2%;
-    margin-left: 46%;
-    gap: 40px;
-    text-decoration: none;
-  }
-
-
-
-
-  .boton
- {
-    align-items: center;
-    justify-content: center;
-    display: inline-f;
-    width: 15%;
-    height: 40px;
-    background-color: rgb(0, 0, 0);
-    color: rgb(156, 16, 16);
-    font-size: 20px;
-    border: none;
-    cursor: pointer;
-    
- 
-  }
-
-   
-   
-    .boton:hover
-    {
-
-     background-color: rgb(165, 42, 128);
-
-    }
-
-  
-
-  
-
-
-    .derechos
-   {
-      text-align: center;
-      margin-top: 7%;
-      margin-left: 0.2%;
-    }
-
-
-   .logohome
-   {
-      margin-bottom: 8%;
-      margin-left: 36%;
-   }
-
-
-   
-
-
-   .contenedorbotonhomes{
-
-    display: grid;
-
-    
-    margin-left: 39%;
-    gap: 40px;
-    text-decoration: none;
-    
-
-
-   }
-
-   .botoneshome{
-
-
-    align-items: center;
-    justify-content: center;
-    width: 34%;
-    height: 80px;
-    background-color: rgb(0, 0, 0);
-    color: rgb(241, 241, 241);
-    font-size: 20px;
-    border: none;
-    cursor: pointer;
-    font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif;
-  }
-  
-
-  
-  .botoneshome:hover
-  {
-
-   background-color: rgb(63, 165, 42);
-
-  }
-
-
-  a{
-
-    text-decoration: none;
-    color: antiquewhite;
-  }
-
-
-  .derechos
-  {
-    margin-top:5%;
-    margin-bottom: 2%;
-    text-align: center;
-    font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif; 
-    line-height:2.5rem;
-  }
-    
-
-
-
-    .seccion1
-    {
-     display: flex;
-     margin-top: 11%;
-     gap: 164px;
-     justify-content: center;
-     text-align: center;
-    }
-
-
-    .seccion2
-    {
-      display: flex;
-      justify-content: center;
-      gap: 164px;
-      margin-top: 6%;
-      margin-bottom: 6%;
-    }
-
-
-    .informacion
-{
-
-  padding: 3%;
-  font-size: 19;
-  color: rgb(10, 10, 10);
-  height: 370px;
-  width: 190px;
-  text-align: center;
-  box-shadow: 0 4px 24px 0 rgb(223, 208, 9);
-  background-color: rgb(230, 240, 238);
-  border-radius: 5%;
-  justify-content: center;
-  line-height:2rem;
-  font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif;
-}
-
-
-.redes
-  {
-
-    margin-top: 7%;
-    justify-content: center;
-    display: flex;
-    gap: 30px;
-  }
-
-.redes2
-{
-  margin-left: 45%;
-  cursor: pointer;
-}
-
-
-
-.botonvolver{
-
-
-
-  align-items: center;
-  justify-content: center;
-  width: 34%;
-  height: 80px;
-  background-color: rgb(0, 0, 0);
-  color: rgb(241, 241, 241);
-  font-size: 20px;
-  border: none;
-  cursor: pointer;
-  font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif;
-  text-align: center;
- 
-}
-
-
-  
-.botonvolver:hover
-{
-
- background-color: rgb(230, 137, 84);
-
-}
-
-
-
-.contenedorbotonvolver{
-
-  display: grid;
-  margin-left: 39%;
-  gap: 40px;
-  text-decoration: none;
-  
-
-
- }
